fix: bug with localstorage
