+++ src/components/index.ts
@@ -3,5 +3,15 @@ import { TableRow } from './tableRow/TableRow';
 import { ButtonPrimary, ButtonSecondary } from './buttons';
 import { Chart } from './chart/Chart';
 import { ModalAdd, ModalPortfolio } from './modals';
+import { Portfolio } from './portfolio/Portfolio';
 
-export { TableHead, TableRow, ButtonPrimary, ButtonSecondary, Chart, ModalAdd, ModalPortfolio };
+export {
+  TableHead,
+  TableRow,
+  ButtonPrimary,
+  ButtonSecondary,
+  Chart,
+  ModalAdd,
+  ModalPortfolio,
+  Portfolio,
+};
